refactor; cleanup
https://github.com/metasfresh/metasfresh-webui-api/issues/1064
