+++ src/main/java/de/metas/ui/web/document/filter/DocumentFilterParamDescriptor.java
@@ -3,14 +3,13 @@ package de.metas.ui.web.document.filter;
 import de.metas.i18n.ITranslatableString;
 import de.metas.i18n.ImmutableTranslatableString;
 import de.metas.ui.web.document.filter.DocumentFilterParam.Operator;
-import de.metas.ui.web.window.descriptor.DocumentFieldDescriptor;
+import de.metas.ui.web.window.datatypes.DataTypes;
 import de.metas.ui.web.window.descriptor.DocumentFieldWidgetType;
 import de.metas.ui.web.window.descriptor.LookupDescriptor;
 import de.metas.ui.web.window.descriptor.factory.standard.DescriptorsFactoryHelper;
 import de.metas.ui.web.window.model.lookup.LookupDataSource;
 import de.metas.ui.web.window.model.lookup.LookupDataSourceFactory;
 import de.metas.util.Check;
-
 import lombok.Value;
 
 /*
@@ -133,7 +132,7 @@ public final class DocumentFilterParamDescriptor
 
 	private Object convertValueToFieldType(final Object value)
 	{
-		return DocumentFieldDescriptor.convertToValueClass(getFieldName(), value, getWidgetType(), getValueClass(), getLookupDataSourceOrNull());
+		return DataTypes.convertToValueClass(getFieldName(), value, getWidgetType(), getValueClass(), getLookupDataSourceOrNull());
 	}
 
 	public boolean isAutoFilter()
