+++ src/main/java/de/metas/ui/web/window/model/lookup/LookupDataSource.java
@@ -77,7 +77,6 @@ public interface LookupDataSource extends LookupValueByIdSupplier
 
 	List<CCacheStats> getCacheStats();
 
-	@Override
 	DocumentZoomIntoInfo getDocumentZoomInto(final int id);
 
 	/** @return optional WindowId to be used when zooming into */
