+++ src/main/java/de/metas/ui/web/window/model/lookup/LookupValueByIdSupplier.java
@@ -15,18 +15,17 @@ import de.metas.ui.web.window.datatypes.LookupValue;
  * 
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
  * 
  * You should have received a copy of the GNU General Public
- * License along with this program.  If not, see
+ * License along with this program. If not, see
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
  * #L%
  */
 
+@FunctionalInterface
 public interface LookupValueByIdSupplier
 {
 	LookupValue findById(Object id);
-	
-	DocumentZoomIntoInfo getDocumentZoomInto(int id);
 }
